requirements.txt
keras==2.8.0
numpy==1.21.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pandas==1.3.5
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools-git==1.2
tensorboard==2.8.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-addons==0.17.1
natsort==5.5.0
Pillow==7.1.2
